+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Lob Upgrade Tech doubles the fire rate and removes the recharge cost
 - Omega Upgrade Tech now also doubles the rate of fire
 - Artemis Upgrade Tech changed to allow planetary targeting
+- Avenger Upgrade Tech changed to add a railgun
 
 ## v5.61.0 - 2024-03-13